rename moh_mappings to mohschema
